refactor(codegen): close out Phase 1A — atoms and text pipeline
Phase 1A Batch 5 (Moves A and B from docs/PHASE1-PLAN.md).  Final
two extractions of Phase 1A.

* Move A — append 9 atom handlers (StringLit, SetLiteral, Subscript,
  Superscript, Tuple, RelationalImage, SequenceLiteral,
  TupleProjection, BagLiteral) to codegen/expressions.py.  These
  are simple expression handlers the original Batch 2 listing
  did not enumerate.
* Move B — extract the entire plain-text-to-LaTeX pipeline into a
  new mixin codegen/text_pipeline.py.  ~25 helpers and two
  ClassVars (_ALLOWED_LATEX_COMMANDS, _BLOCKED_LATEX_COMMANDS) that
  implement TEXT: / PURETEXT: / LATEX: emission and the
  inline-text portions of Part rendering.

Method bodies are byte-identical to their counterparts in the
pre-refactor monolithic latex_gen.py.

The CodegenDispatch TYPE_CHECKING block grows by one entry
(_dollar_sanitise_registry) so the moved _pre_sanitise_dollars /
_restore_dollar_sanitise see their shared state.

Gate 5 verified:
* make check — clean.
* make test-e2e — 159/159 byte-identical.
* make refactor-diff — 190/190 byte-identical vs main.

latex_gen.py shrinks from 2514 → 598 lines.  Cumulative reduction
from the original 6864-line monolith: 91.3%.  Phase 1A complete.
