+++ src/txt2tex/codegen/__init__.py
@@ -14,5 +14,6 @@ from txt2tex.codegen import (
     proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     text_blocks,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    text_pipeline,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     types,  # noqa: F401  # pyright: ignore[reportUnusedImport]
 )
